# docs/swagger.yaml
definitions:
...
  response.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
  response.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
  socialmedia.SocialMediaCreate:
    properties:
      media_url:
        type: string
      name:
        type: string
      user_id:
        type: integer
    required:
    - media_url
    - name
    type: object
  socialmedia.SocialMediaRes:
    properties:
      id:
        type: integer
      media_url:
        type: string
      name:
        type: string
      user_id:
        type: integer
    type: object
  socialmedia.SocialMediaUpdate:
    properties:
      id:
        type: integer
      media_url:
        type: string
      name:
        type: string
      user_id:
        type: integer
    required:
    - media_url
    - name
    type: object
  token.Tokens:
    properties:
      access_token:
        type: string
      id_token:
        type: string
      refresh_token:
        type: string
    type: object
  user.UserCreate:
    properties:
      age:
        minimum: 9
        type: integer
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - age
    - email
    - password
    - username
    type: object
  user.UserLogin:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: user payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/token.Tokens'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login
  /register:
    post:
      consumes:
      - application/json
      description: Registration
      parameters:
      - description: user payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.UserCreate'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Register
...
  /user/media:
    get:
      consumes:
      - application/json
      description: FindAll  Social Media
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/socialmedia.SocialMediaRes'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: FindAll  Media
    post:
      consumes:
      - application/json
      description: Insert Social Media
      parameters:
      - description: media payload
        in: body
        name: media
        required: true
        schema:
          $ref: '#/definitions/socialmedia.SocialMediaCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Insert Media
  /user/media/{idMedia}:
    delete:
      consumes:
      - application/json
      description: Delete  Social Media
      parameters:
      - description: idMedia
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete  Media
    get:
      consumes:
      - application/json
      description: FindBy Id  Social Media
      parameters:
      - description: idMedia
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/socialmedia.SocialMediaRes'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: FindBy Id Media
    put:
      consumes:
      - application/json
      description: Update Social Media
      parameters:
      - description: media payload
        in: body
        name: media
        required: true
        schema:
          $ref: '#/definitions/socialmedia.SocialMediaUpdate'
      - description: idMedia
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update Media
...
swagger: "2.0"